+++ Vungle/CHANGELOG.md
@@ -1,4 +1,8 @@
 ## Changelog
+* 6.5.2.0
+   * This version of the adapters has been certified with Vungle 6.5.2.
+   * Add support for Vungle's newly-introduced banner format.
+
   * 6.4.6.1
     * Remove duplicate `rewardedVideoDidAppearForCustomEvent` callback.
     * Adapter logging cleanup.